+++ ui/component/creatorAnalytics/view.jsx
@@ -27,7 +27,9 @@ export default function CreatorAnalytics(props: Props) {
   const [error, setError] = React.useState();
   const [fetchingStats, setFetchingStats] = React.useState(false);
   const claimId = claim && claim.claim_id;
-  const channelHasClaims = claim && claim.meta && claim.meta.claims_in_channel && claim.meta.claims_in_channel > 0;
+  // TODO: put this back when hubs are fixed
+  // const channelHasClaims = claim && claim.meta && claim.meta.claims_in_channel && claim.meta.claims_in_channel > 0 || true;
+  const channelHasClaims = true;
 
   React.useEffect(() => {
     setStats(null);