+++ ui/component/downloadProgress/index.js
@@ -1,16 +1,24 @@
 import { connect } from 'react-redux';
 import DownloadProgress from './view';
-import { doSetPlayingUri, doStopDownload } from 'redux/actions/content';
+import { doSetPlayingUri, doStopDownload, doUpdateDownloadingStatus } from 'redux/actions/content';
+import { selectFileInfosByOutpoint } from 'lbry-redux';
+
 const select = (state) => {
-  // console.log('DownloadProgress select state', state.fileInfo);
+  const byOutpoint = selectFileInfosByOutpoint(state);
+  const runningByOutpoint = [];
+
+  for (const key in byOutpoint) {
+    if (byOutpoint[key] && byOutpoint[key].status === 'running') runningByOutpoint.push(byOutpoint[key]);
+  }
 
   return {
-    downloadList: state.fileInfo.byOutpoint,
+    downloadList: runningByOutpoint,
   };
 };
 
 const perform = (dispatch) => ({
   pause: () => dispatch(doSetPlayingUri({ uri: null })),
-  stopDownload: (outpoint, sd_hash) => dispatch(doStopDownload(outpoint, sd_hash)),
+  updateDownloadingStatus: (outpoint) => dispatch(doUpdateDownloadingStatus(outpoint)),
+  stopDownload: (outpoint) => dispatch(doStopDownload(outpoint)),
 });
 export default connect(select, perform)(DownloadProgress);