+++ ui/redux/selectors/file_info.js
@@ -1,11 +1,11 @@
 import { selectClaimsByUri, selectIsFetchingClaimListMine, selectMyClaims } from 'lbry-redux';
 import { createSelector } from 'reselect';
 
-export const selectState = state => state.fileInfo || {};
+export const selectState = (state) => state.fileInfo || {};
 
-export const selectFileInfosByOutpoint = createSelector(selectState, state => state.byOutpoint || {});
+export const selectFileInfosByOutpoint = createSelector(selectState, (state) => state.byOutpoint || {});
 
-export const selectIsFetchingFileList = createSelector(selectState, state => state.isFetchingFileList);
+export const selectIsFetchingFileList = createSelector(selectState, (state) => state.isFetchingFileList);
 
 export const selectIsFetchingFileListDownloadedOrPublished = createSelector(
   selectIsFetchingFileList,
@@ -13,31 +13,31 @@ export const selectIsFetchingFileListDownloadedOrPublished = createSelector(
   (isFetchingFileList, isFetchingClaimListMine) => isFetchingFileList || isFetchingClaimListMine
 );
 
-export const makeSelectFileInfoForUri = uri =>
+export const makeSelectFileInfoForUri = (uri) =>
   createSelector(selectClaimsByUri, selectFileInfosByOutpoint, (claims, byOutpoint) => {
     const claim = claims[uri];
     const outpoint = claim ? `${claim.txid}:${claim.nout}` : undefined;
     return outpoint ? byOutpoint[outpoint] : undefined;
   });
 
-export const selectDownloadingByOutpoint = createSelector(selectState, state => state.downloadingByOutpoint || {});
+export const selectDownloadingByOutpoint = createSelector(selectState, (state) => state.downloadingByOutpoint || {});
 
-export const makeSelectDownloadingForUri = uri =>
+export const makeSelectDownloadingForUri = (uri) =>
   createSelector(selectDownloadingByOutpoint, makeSelectFileInfoForUri(uri), (byOutpoint, fileInfo) => {
     if (!fileInfo) return false;
     return byOutpoint[fileInfo.outpoint];
   });
 
-export const selectUrisLoading = createSelector(selectState, state => state.urisLoading || {});
+export const selectUrisLoading = createSelector(selectState, (state) => state.urisLoading || {});
 
-export const makeSelectLoadingForUri = uri => createSelector(selectUrisLoading, byUri => byUri && byUri[uri]);
+export const makeSelectLoadingForUri = (uri) => createSelector(selectUrisLoading, (byUri) => byUri && byUri[uri]);
 
 export const selectFileInfosDownloaded = createSelector(
   selectFileInfosByOutpoint,
   selectMyClaims,
   (byOutpoint, myClaims) =>
-    Object.values(byOutpoint).filter(fileInfo => {
-      const myClaimIds = myClaims.map(claim => claim.claim_id);
+    Object.values(byOutpoint).filter((fileInfo) => {
+      const myClaimIds = myClaims.map((claim) => claim.claim_id);
 
       return fileInfo && myClaimIds.indexOf(fileInfo.claim_id) === -1 && (fileInfo.completed || fileInfo.written_bytes);
     })
@@ -59,7 +59,7 @@ export const selectDownloadingFileInfos = createSelector(
     const outpoints = Object.keys(downloadingByOutpoint);
     const fileInfos = [];
 
-    outpoints.forEach(outpoint => {
+    outpoints.forEach((outpoint) => {
       const fileInfo = fileInfosByOutpoint[outpoint];
 
       if (fileInfo) fileInfos.push(fileInfo);
@@ -69,10 +69,10 @@ export const selectDownloadingFileInfos = createSelector(
   }
 );
 
-export const selectTotalDownloadProgress = createSelector(selectDownloadingFileInfos, fileInfos => {
+export const selectTotalDownloadProgress = createSelector(selectDownloadingFileInfos, (fileInfos) => {
   const progress = [];
 
-  fileInfos.forEach(fileInfo => {
+  fileInfos.forEach((fileInfo) => {
     progress.push((fileInfo.written_bytes / fileInfo.total_bytes) * 100);
   });
 
@@ -82,4 +82,4 @@ export const selectTotalDownloadProgress = createSelector(selectDownloadingFileI
   return -1;
 });
 
-export const selectFileInfoErrors = createSelector(selectState, state => state.errors || {});
+export const selectFileInfoErrors = createSelector(selectState, (state) => state.errors || {});